backend/requirements.txt
asgiref==3.10.0
Django==4.2.16
django-auditlog==3.3.0
django-cors-headers==4.9.0
django-dbbackup==5.0.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
rest-framework-simplejwt==0.0.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
